project/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
sqlparse==0.5.1
tzdata==2024.1
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
